refactor: formattable.ts into formattable/, including Formattable.from() -> asFormattable()
